Add files via upload
A tool which when runs on the victim's system, it would start sending data back to the attacker on the attacker's email. 
The tool should support inputs from Keyboard, Mouse, Screenshots and Mic Inputs. 

What makes this tool different from other spyware is that if the victim tries to read the code and the email of the attacker, the program will self-destruct itself. 

The tool should use public Python libraries which should result in a positive virustotal score meaning that no antivirus should be able to scan the program as malicious.